+++ .include/controller/Commitdiff.php
@@ -1,32 +1,25 @@
 <?php
+namespace GitPHP\Controller;
 
-class GitPHP_Controller_Commitdiff extends GitPHP_Controller_DiffBase
+class Commitdiff extends DiffBase
 {
-    /**
-	 * __construct
-	 *
-	 * Constructor
-	 *
-	 * @access public
-	 * @return controller
-	 */
-	    public function __construct()
+    public function __construct()
     {
         parent::__construct();
         if (!$this->project) {
-            throw new GitPHP_MessageException(__('Project is required'), true);
+            throw new \GitPHP_MessageException(__('Project is required'), true);
         }
     }
 
     /**
-	 * GetTemplate
-	 *
-	 * Gets the template for this controller
-	 *
-	 * @access protected
-	 * @return string template filename
-	 */
-	    protected function GetTemplate()
+     * GetTemplate
+     *
+     * Gets the template for this controller
+     *
+     * @access protected
+     * @return string template filename
+     */
+    protected function GetTemplate()
     {
         if (isset($this->params['plain']) && ($this->params['plain'] === true)) {
             return 'commitdiffplain.tpl';
@@ -35,14 +28,14 @@ class GitPHP_Controller_Commitdiff extends GitPHP_Controller_DiffBase
     }
 
     /**
-	 * GetCacheKey
-	 *
-	 * Gets the cache key for this controller
-	 *
-	 * @access protected
-	 * @return string cache key
-	 */
-	    protected function GetCacheKey()
+     * GetCacheKey
+     *
+     * Gets the cache key for this controller
+     *
+     * @access protected
+     * @return string cache key
+     */
+    protected function GetCacheKey()
     {
         $key = (isset($this->params['hash']) ? $this->params['hash'] : '')
             . '|' . (isset($this->params['hashparent']) ? $this->params['hashparent'] : '')
@@ -52,15 +45,15 @@ class GitPHP_Controller_Commitdiff extends GitPHP_Controller_DiffBase
     }
 
     /**
-	 * GetName
-	 *
-	 * Gets the name of this controller's action
-	 *
-	 * @access public
-	 * @param boolean $local true if caller wants the localized action name
-	 * @return string action name
-	 */
-	    public function GetName($local = false)
+     * GetName
+     *
+     * Gets the name of this controller's action
+     *
+     * @access public
+     * @param boolean $local true if caller wants the localized action name
+     * @return string action name
+     */
+    public function GetName($local = false)
     {
         if ($local) {
             return __('commitdiff');
@@ -69,13 +62,13 @@ class GitPHP_Controller_Commitdiff extends GitPHP_Controller_DiffBase
     }
 
     /**
-	 * ReadQuery
-	 *
-	 * Read query into parameters
-	 *
-	 * @access protected
-	 */
-	    protected function ReadQuery()
+     * ReadQuery
+     *
+     * Read query into parameters
+     *
+     * @access protected
+     */
+    protected function ReadQuery()
     {
         parent::ReadQuery();
 
@@ -92,13 +85,13 @@ class GitPHP_Controller_Commitdiff extends GitPHP_Controller_DiffBase
     }
 
     /**
-	 * LoadHeaders
-	 *
-	 * Loads headers for this template
-	 *
-	 * @access protected
-	 */
-	    protected function LoadHeaders()
+     * LoadHeaders
+     *
+     * Loads headers for this template
+     *
+     * @access protected
+     */
+    protected function LoadHeaders()
     {
         parent::LoadHeaders();
 
@@ -108,13 +101,13 @@ class GitPHP_Controller_Commitdiff extends GitPHP_Controller_DiffBase
     }
 
     /**
-	 * LoadData
-	 *
-	 * Loads data for this template
-	 *
-	 * @access protected
-	 */
-	    protected function LoadData()
+     * LoadData
+     *
+     * Loads data for this template
+     *
+     * @access protected
+     */
+    protected function LoadData()
     {
         parent::LoadData();
         $co = $this->project->GetCommit($this->params['hash']);
@@ -129,12 +122,12 @@ class GitPHP_Controller_Commitdiff extends GitPHP_Controller_DiffBase
         }
 
         $renames = true;
-        $DiffContext = new DiffContext();
+        $DiffContext = new \DiffContext();
         $DiffContext->setRenames($renames)
             ->setContext($this->params['context'])
             ->setIgnoreWhitespace($this->params['ignorewhitespace'])
             ->setIgnoreFormatting($this->params['ignoreformat']);
-        $treediff = new GitPHP_TreeDiff(
+        $treediff = new \GitPHP_TreeDiff(
             $this->project,
             $this->params['hash'],
             (isset($this->params['hashparent']) ? $this->params['hashparent'] : ''),
@@ -144,7 +137,7 @@ class GitPHP_Controller_Commitdiff extends GitPHP_Controller_DiffBase
         $this->loadReviewsLinks($co, implode('', $co->GetComment()));
 
         if (empty($this->params['sidebyside'])) {
-            include_once(GitPHP_Util::AddSlash('lib/syntaxhighlighter') . "syntaxhighlighter.php");
+            include_once(\GitPHP_Util::AddSlash('lib/syntaxhighlighter') . "syntaxhighlighter.php");
             $this->tpl->assign('sexy', 1);
             $this->tpl->assign('highlighter_no_ruler', 1);
             $this->tpl->assign('highlighter_diff_enabled', 1);
@@ -153,19 +146,21 @@ class GitPHP_Controller_Commitdiff extends GitPHP_Controller_DiffBase
             $extensions = [];
             $statuses = [];
             $folders = [];
-            foreach($treediff as $filediff) {
+            foreach ($treediff as $filediff) {
                 /** @var \GitPHP_FileDiff $filediff */
 
                 $extensions[$filediff->getToFileExtension()] = $filediff->getToFileExtension();
                 $statuses[$filediff->GetStatus()] = $filediff->GetStatus();
                 $folders[$filediff->getToFileRootFolder()] = $filediff->getToFileRootFolder();
 
-                $SH = new SyntaxHighlighter($filediff->GetToFile());
+                $SH = new \SyntaxHighlighter($filediff->GetToFile());
                 $brashes = array_merge($SH->getBrushesList(), $brashes);
-                $filediff->SetDecorationData([
-                    'highlighter_brushes' => $SH->getBrushesList(),
-                    'highlighter_brush_name' => $SH->getBrushName(),
-                ]);
+                $filediff->SetDecorationData(
+                    [
+                        'highlighter_brushes' => $SH->getBrushesList(),
+                        'highlighter_brush_name' => $SH->getBrushName(),
+                    ]
+                );
                 $this->tpl->assign('extracss_files', $SH->getCssList());
                 $this->tpl->assign('extrajs_files', $SH->getJsList());
             }
