+++ .include/controller/Tags.php
@@ -1,0 +1,83 @@
+<?php
+namespace GitPHP\Controller;
+
+class Tags extends Base
+{
+    public function __construct()
+    {
+        parent::__construct();
+        if (!$this->project) {
+            throw new \GitPHP_MessageException(__('Project is required'), true);
+        }
+    }
+
+    /**
+     * GetTemplate
+     *
+     * Gets the template for this controller
+     *
+     * @access protected
+     * @return string template filename
+     */
+    protected function GetTemplate()
+    {
+        return 'tags.tpl';
+    }
+
+    /**
+     * GetCacheKey
+     *
+     * Gets the cache key for this controller
+     *
+     * @access protected
+     * @return string cache key
+     */
+    protected function GetCacheKey()
+    {
+        return '';
+    }
+
+    /**
+     * GetName
+     *
+     * Gets the name of this controller's action
+     *
+     * @access public
+     * @param boolean $local true if caller wants the localized action name
+     * @return string action name
+     */
+    public function GetName($local = false)
+    {
+        if ($local) {
+            return __('tags');
+        }
+        return 'tags';
+    }
+
+    /**
+     * ReadQuery
+     *
+     * Read query into parameters
+     *
+     * @access protected
+     */
+    protected function ReadQuery() {}
+
+    /**
+     * LoadData
+     *
+     * Loads data for this template
+     *
+     * @access protected
+     */
+    protected function LoadData()
+    {
+        $head = $this->project->GetHeadCommit();
+        $this->tpl->assign("head", $head);
+
+        $taglist = $this->project->GetTags();
+        if (isset($taglist) && (count($taglist) > 0)) {
+            $this->tpl->assign("taglist", $taglist);
+        }
+    }
+}
