+++ autoload.php
@@ -1,81 +1,79 @@
 <?php
 return array(
-    'GitPHP\\Error'                         => GITPHP_INCLUDEDIR . 'Error.php',
-    'GitPHP_Application'                    => GITPHP_INCLUDEDIR . 'Application.php',
-    'GitPHP_Db'                             => GITPHP_INCLUDEDIR . 'Db.class.php',
-    'GitPHP_Config'                         => GITPHP_INCLUDEDIR . 'Config.class.php',
-    'GitPHP_Log'                            => GITPHP_INCLUDEDIR . 'Log.class.php',
-    'GitPHP_MessageException'               => GITPHP_INCLUDEDIR . 'MessageException.class.php',
-    'GitPHP_Resource'                       => GITPHP_INCLUDEDIR . 'Resource.class.php',
-    'GitPHP_Util'                           => GITPHP_INCLUDEDIR . 'Util.class.php',
-    'GitPHP_User'                           => GITPHP_INCLUDEDIR . 'User.class.php',
-    'GitPHP_Session'                        => GITPHP_INCLUDEDIR . 'Session.class.php',
-    \GitPHP\Jira::class                     => GITPHP_INCLUDEDIR . 'Jira.php',
-    \GitPHP\Redmine::class                  => GITPHP_INCLUDEDIR . 'Redmine.php',
-    \GitPHP\JiraRestClient::class           => GITPHP_INCLUDEDIR . 'JiraRestClient.php',
-    \GitPHP\RedmineRestClient::class        => GITPHP_INCLUDEDIR . 'RedmineRestClient.php',
-    \GitPHP\Acl::class                      => GITPHP_INCLUDEDIR . 'Acl.php',
-    \GitPHP\Db_Result::class                => GITPHP_INCLUDEDIR . 'Result.php',
-    \GitPHP\Http_Response::class            => GITPHP_INCLUDEDIR . 'Http_Response.php',
-    \GitPHP\Tracker::class                  => GITPHP_INCLUDEDIR . 'Tracker.class.php',
-    'Model_Gitosis'                         => GITPHP_INCLUDEDIR . 'Model_Gitosis.php',
-    'GitPHP_Gitosis'                        => GITPHP_INCLUDEDIR . 'Gitosis.class.php',
-    'GitPHP_Cache'                          => GITPHP_CACHEDIR   . 'Cache.class.php',
-    'GitPHP_Archive'                        => GITPHP_GITOBJECTDIR . 'Archive.class.php',
-    'GitPHP_Blob'                           => GITPHP_GITOBJECTDIR . 'Blob.class.php',
-    'GitPHP_BranchDiff'                     => GITPHP_GITOBJECTDIR . 'BranchDiff.class.php',
-    'GitPHP_Commit'                         => GITPHP_GITOBJECTDIR . 'Commit.class.php',
-    'GitPHP_DiffExe'                        => GITPHP_GITOBJECTDIR . 'DiffExe.class.php',
-    'GitPHP_FileDiff'                       => GITPHP_GITOBJECTDIR . 'FileDiff.class.php',
-    'GitPHP_FilesystemObject'               => GITPHP_GITOBJECTDIR . 'FilesystemObject.class.php',
-    'GitPHP_GitExe'                         => GITPHP_GITOBJECTDIR . 'GitExe.class.php',
-    'GitPHP_GitObject'                      => GITPHP_GITOBJECTDIR . 'GitObject.class.php',
-    'GitPHP_Head'                           => GITPHP_GITOBJECTDIR . 'Head.class.php',
-    'GitPHP_Project'                        => GITPHP_GITOBJECTDIR . 'Project.class.php',
-    'GitPHP_ProjectList'                    => GITPHP_GITOBJECTDIR . 'ProjectList.class.php',
-    'GitPHP_ProjectListArray'               => GITPHP_GITOBJECTDIR . 'ProjectListArray.class.php',
-    'GitPHP_ProjectListArrayLegacy'         => GITPHP_GITOBJECTDIR . 'ProjectListArrayLegacy.class.php',
-    'GitPHP_ProjectListBase'                => GITPHP_GITOBJECTDIR . 'ProjectListBase.class.php',
-    'GitPHP_ProjectListDirectory'           => GITPHP_GITOBJECTDIR . 'ProjectListDirectory.class.php',
-    'GitPHP_ProjectListFile'                => GITPHP_GITOBJECTDIR . 'ProjectListFile.class.php',
-    'GitPHP_Ref'                            => GITPHP_GITOBJECTDIR . 'Ref.class.php',
-    'GitPHP_Tag'                            => GITPHP_GITOBJECTDIR . 'Tag.class.php',
-    'GitPHP_TmpDir'                         => GITPHP_GITOBJECTDIR . 'TmpDir.class.php',
-    'GitPHP_Tree'                           => GITPHP_GITOBJECTDIR . 'Tree.class.php',
-    'GitPHP_TreeDiff'                       => GITPHP_GITOBJECTDIR . 'TreeDiff.class.php',
-    'DiffHighlighter'                       => GITPHP_GITOBJECTDIR . 'DiffHighlighter.php',
-    'DiffContext'                           => GITPHP_GITOBJECTDIR . 'DiffContext.php',
-    'GitPHP_ControllerBase'                 => GITPHP_CONTROLLERDIR . 'ControllerBase.class.php',
-    'GitPHP_Controller_Blame'               => GITPHP_CONTROLLERDIR . 'Controller_Blame.class.php',
-    'GitPHP_Controller_Blob'                => GITPHP_CONTROLLERDIR . 'Controller_Blob.class.php',
-    'GitPHP_Controller_Blobdiff'            => GITPHP_CONTROLLERDIR . 'Controller_Blobdiff.class.php',
-    'GitPHP_Controller_Branchdiff'          => GITPHP_CONTROLLERDIR . 'Controller_Branchdiff.class.php',
-    'GitPHP_Controller_CheckSession'        => GITPHP_CONTROLLERDIR . 'Controller_CheckSession.php',
-    'GitPHP_Controller_Comment'             => GITPHP_CONTROLLERDIR . 'Controller_Comment.class.php',
-    'GitPHP_Controller_Commit'              => GITPHP_CONTROLLERDIR . 'Controller_Commit.class.php',
-    'GitPHP_Controller_Commitdiff'          => GITPHP_CONTROLLERDIR . 'Controller_Commitdiff.class.php',
-    'GitPHP_Controller_DiffBase'            => GITPHP_CONTROLLERDIR . 'Controller_DiffBase.class.php',
-    'GitPHP_Controller_Feed'                => GITPHP_CONTROLLERDIR . 'Controller_Feed.class.php',
-    'GitPHP_Controller_GitosisAccess'       => GITPHP_CONTROLLERDIR . 'Controller_GitosisAccess.class.php',
-    'GitPHP_Controller_GitosisApply'        => GITPHP_CONTROLLERDIR . 'Controller_GitosisApply.class.php',
-    'GitPHP_Controller_GitosisBase'         => GITPHP_CONTROLLERDIR . 'Controller_GitosisBase.class.php',
-    'GitPHP_Controller_GitosisRepositories' => GITPHP_CONTROLLERDIR . 'Controller_GitosisRepositories.class.php',
-    'GitPHP_Controller_GitosisUsers'        => GITPHP_CONTROLLERDIR . 'Controller_GitosisUsers.class.php',
-    'GitPHP_Controller_Heads'               => GITPHP_CONTROLLERDIR . 'Controller_Heads.class.php',
-    'GitPHP_Controller_History'             => GITPHP_CONTROLLERDIR . 'Controller_History.class.php',
-    'GitPHP_Controller_Log'                 => GITPHP_CONTROLLERDIR . 'Controller_Log.class.php',
-    'GitPHP_Controller_Login'               => GITPHP_CONTROLLERDIR . 'Controller_Login.class.php',
-    'GitPHP_Controller_Logout'              => GITPHP_CONTROLLERDIR . 'Controller_Logout.class.php',
-    'GitPHP_Controller_Message'             => GITPHP_CONTROLLERDIR . 'Controller_Message.class.php',
-    'GitPHP_Controller_Review'              => GITPHP_CONTROLLERDIR . 'Controller_Review.php',
-    'GitPHP_Controller_Search'              => GITPHP_CONTROLLERDIR . 'Controller_Search.class.php',
-    'GitPHP_Controller_Snapshot'            => GITPHP_CONTROLLERDIR . 'Controller_Snapshot.class.php',
-    'GitPHP_Controller_Tag'                 => GITPHP_CONTROLLERDIR . 'Controller_Tag.class.php',
-    'GitPHP_Controller_Tags'                => GITPHP_CONTROLLERDIR . 'Controller_Tags.class.php',
-    'GitPHP_Controller_Tree'                => GITPHP_CONTROLLERDIR . 'Controller_Tree.class.php',
-    'GitPHP_Controller_Project'             => GITPHP_CONTROLLERDIR . 'Controller_Project.class.php',
-    'GitPHP_Controller_ProjectList'         => GITPHP_CONTROLLERDIR . 'Controller_ProjectList.class.php',
-    'GitPHP_Controller_Git'                 => GITPHP_CONTROLLERDIR . 'Controller_Git.php',
-    'GitPHP_Controller_SearchText'          => GITPHP_CONTROLLERDIR . 'Controller_SearchText.class.php'
-
+    'GitPHP\\Error'                               => GITPHP_INCLUDEDIR . 'Error.php',
+    'GitPHP_Application'                          => GITPHP_INCLUDEDIR . 'Application.php',
+    'GitPHP_Db'                                   => GITPHP_INCLUDEDIR . 'Db.class.php',
+    'GitPHP_Config'                               => GITPHP_INCLUDEDIR . 'Config.class.php',
+    'GitPHP_Log'                                  => GITPHP_INCLUDEDIR . 'Log.class.php',
+    'GitPHP_MessageException'                     => GITPHP_INCLUDEDIR . 'MessageException.class.php',
+    'GitPHP_Resource'                             => GITPHP_INCLUDEDIR . 'Resource.class.php',
+    'GitPHP_Util'                                 => GITPHP_INCLUDEDIR . 'Util.class.php',
+    'GitPHP_User'                                 => GITPHP_INCLUDEDIR . 'User.class.php',
+    'GitPHP_Session'                              => GITPHP_INCLUDEDIR . 'Session.class.php',
+    \GitPHP\Jira::class                           => GITPHP_INCLUDEDIR . 'Jira.php',
+    \GitPHP\Redmine::class                        => GITPHP_INCLUDEDIR . 'Redmine.php',
+    \GitPHP\JiraRestClient::class                 => GITPHP_INCLUDEDIR . 'JiraRestClient.php',
+    \GitPHP\RedmineRestClient::class              => GITPHP_INCLUDEDIR . 'RedmineRestClient.php',
+    \GitPHP\Acl::class                            => GITPHP_INCLUDEDIR . 'Acl.php',
+    \GitPHP\Db_Result::class                      => GITPHP_INCLUDEDIR . 'Result.php',
+    \GitPHP\Http_Response::class                  => GITPHP_INCLUDEDIR . 'Http_Response.php',
+    \GitPHP\Tracker::class                        => GITPHP_INCLUDEDIR . 'Tracker.class.php',
+    'Model_Gitosis'                               => GITPHP_INCLUDEDIR . 'Model_Gitosis.php',
+    'GitPHP_Gitosis'                              => GITPHP_INCLUDEDIR . 'Gitosis.class.php',
+    'GitPHP_Cache'                                => GITPHP_CACHEDIR . 'Cache.class.php',
+    'GitPHP_Archive'                              => GITPHP_GITOBJECTDIR . 'Archive.class.php',
+    'GitPHP_Blob'                                 => GITPHP_GITOBJECTDIR . 'Blob.class.php',
+    'GitPHP_BranchDiff'                           => GITPHP_GITOBJECTDIR . 'BranchDiff.class.php',
+    'GitPHP_Commit'                               => GITPHP_GITOBJECTDIR . 'Commit.class.php',
+    'GitPHP_DiffExe'                              => GITPHP_GITOBJECTDIR . 'DiffExe.class.php',
+    'GitPHP_FileDiff'                             => GITPHP_GITOBJECTDIR . 'FileDiff.class.php',
+    'GitPHP_FilesystemObject'                     => GITPHP_GITOBJECTDIR . 'FilesystemObject.class.php',
+    'GitPHP_GitExe'                               => GITPHP_GITOBJECTDIR . 'GitExe.class.php',
+    'GitPHP_GitObject'                            => GITPHP_GITOBJECTDIR . 'GitObject.class.php',
+    'GitPHP_Head'                                 => GITPHP_GITOBJECTDIR . 'Head.class.php',
+    'GitPHP_Project'                              => GITPHP_GITOBJECTDIR . 'Project.class.php',
+    'GitPHP_ProjectList'                          => GITPHP_GITOBJECTDIR . 'ProjectList.class.php',
+    'GitPHP_ProjectListArray'                     => GITPHP_GITOBJECTDIR . 'ProjectListArray.class.php',
+    'GitPHP_ProjectListArrayLegacy'               => GITPHP_GITOBJECTDIR . 'ProjectListArrayLegacy.class.php',
+    'GitPHP_ProjectListBase'                      => GITPHP_GITOBJECTDIR . 'ProjectListBase.class.php',
+    'GitPHP_ProjectListDirectory'                 => GITPHP_GITOBJECTDIR . 'ProjectListDirectory.class.php',
+    'GitPHP_ProjectListFile'                      => GITPHP_GITOBJECTDIR . 'ProjectListFile.class.php',
+    'GitPHP_Ref'                                  => GITPHP_GITOBJECTDIR . 'Ref.class.php',
+    'GitPHP_Tag'                                  => GITPHP_GITOBJECTDIR . 'Tag.class.php',
+    'GitPHP_TmpDir'                               => GITPHP_GITOBJECTDIR . 'TmpDir.class.php',
+    'GitPHP_Tree'                                 => GITPHP_GITOBJECTDIR . 'Tree.class.php',
+    'GitPHP_TreeDiff'                             => GITPHP_GITOBJECTDIR . 'TreeDiff.class.php',
+    'DiffHighlighter'                             => GITPHP_GITOBJECTDIR . 'DiffHighlighter.php',
+    'DiffContext'                                 => GITPHP_GITOBJECTDIR . 'DiffContext.php',
+    \GitPHP\Controller\Base::class                => GITPHP_CONTROLLERDIR . 'Base.php',
+    \GitPHP\Controller\Blame::class               => GITPHP_CONTROLLERDIR . 'Blame.php',
+    \GitPHP\Controller\Blob::class                => GITPHP_CONTROLLERDIR . 'Blob.php',
+    \GitPHP\Controller\Blobdiff::class            => GITPHP_CONTROLLERDIR . 'Blobdiff.php',
+    \GitPHP\Controller\Branchdiff::class          => GITPHP_CONTROLLERDIR . 'Branchdiff.php',
+    \GitPHP\Controller\CheckSession::class        => GITPHP_CONTROLLERDIR . 'CheckSession.php',
+    \GitPHP\Controller\Comment::class             => GITPHP_CONTROLLERDIR . 'Comment.php',
+    \GitPHP\Controller\Commit::class              => GITPHP_CONTROLLERDIR . 'Commit.php',
+    \GitPHP\Controller\Commitdiff::class          => GITPHP_CONTROLLERDIR . 'Commitdiff.php',
+    \GitPHP\Controller\DiffBase::class            => GITPHP_CONTROLLERDIR . 'DiffBase.php',
+    \GitPHP\Controller\Feed::class                => GITPHP_CONTROLLERDIR . 'Feed.php',
+    \GitPHP\Controller\GitosisAccess::class       => GITPHP_CONTROLLERDIR . 'GitosisAccess.php',
+    \GitPHP\Controller\GitosisBase::class         => GITPHP_CONTROLLERDIR . 'GitosisBase.php',
+    \GitPHP\Controller\GitosisRepositories::class => GITPHP_CONTROLLERDIR . 'GitosisRepositories.php',
+    \GitPHP\Controller\GitosisUsers::class        => GITPHP_CONTROLLERDIR . 'GitosisUsers.php',
+    \GitPHP\Controller\Heads::class               => GITPHP_CONTROLLERDIR . 'Heads.php',
+    \GitPHP\Controller\History::class             => GITPHP_CONTROLLERDIR . 'History.php',
+    \GitPHP\Controller\Log::class                 => GITPHP_CONTROLLERDIR . 'Log.php',
+    \GitPHP\Controller\Login::class               => GITPHP_CONTROLLERDIR . 'Login.php',
+    \GitPHP\Controller\Logout::class              => GITPHP_CONTROLLERDIR . 'Logout.php',
+    \GitPHP\Controller\Message::class             => GITPHP_CONTROLLERDIR . 'Message.php',
+    \GitPHP\Controller\Review::class              => GITPHP_CONTROLLERDIR . 'Review.php',
+    \GitPHP\Controller\Search::class              => GITPHP_CONTROLLERDIR . 'Search.php',
+    \GitPHP\Controller\Snapshot::class            => GITPHP_CONTROLLERDIR . 'Snapshot.php',
+    \GitPHP\Controller\Tag::class                 => GITPHP_CONTROLLERDIR . 'Tag.php',
+    \GitPHP\Controller\Tags::class                => GITPHP_CONTROLLERDIR . 'Tags.php',
+    \GitPHP\Controller\Tree::class                => GITPHP_CONTROLLERDIR . 'Tree.php',
+    \GitPHP\Controller\Project::class             => GITPHP_CONTROLLERDIR . 'Project.php',
+    \GitPHP\Controller\ProjectList::class         => GITPHP_CONTROLLERDIR . 'ProjectList.php',
+    \GitPHP\Controller\Git::class                 => GITPHP_CONTROLLERDIR . 'Git.php',
+    \GitPHP\Controller\SearchText::class          => GITPHP_CONTROLLERDIR . 'SearchText.php'
 );
